[General] rename folder from types/ to utils/ and extract out constants
