+++ components/EventCard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from 'next/image';
 import { convertToDate, getTime } from '@/utils/utils';
-import { EventType } from '@/types/types';
+import { EventType } from '@/utils/types';
 import {
   AddressContainer,
   AddressIcon,
