+++ jest.config.js
@@ -17,8 +17,6 @@ const customJestConfig = {
 
     '^@/lib/(.*)$': '<rootDir>/lib/$1',
 
-    '^@/types/(.*)$': '<rootDir>/types/$1',
-
     '^@/utils/(.*)$': '<rootDir>/utils/$1',
 
     '@/(.*)$': '<rootDir>/$1',
