+++ utils/types.ts
@@ -27,3 +27,16 @@ export interface RegisterFormFunctions {
   handleLeftArrow: (data: any) => void;
   handleRightArrow: () => void;
 }
+
+/**
+ * Icon Param container.
+ * Used to create icons through iteration
+ * @defaultSrc src of the icon when not selected or hovered
+ * @hoveredsrc src of the icon when selected or hovered
+ * @linkTo where the link of icon directs to
+ */
+export interface IconParams {
+  defaultSrc: string;
+  hoveredsrc: string;
+  linkTo: string;
+}
